requirements.txt
# Core RAG Dependencies
langchain>=0.1.0
streamlit>=1.28.0
openai>=1.3.0
faiss-cpu>=1.7.4

# Document Processing
wikipedia>=1.4.0

# ML and Vector Operations
scikit-learn>=1.3.0
numpy>=1.24.0

# Additional utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Optional: For enhanced text processing
beautifulsoup4>=4.12.0
requests>=2.31.0